+++ js/data/game.js
@@ -1,8 +1,21 @@
 import {
   Settings,
   AnswerValue,
-  AnswerPoint
+  AnswerPoint,
+  testGame,
+  GameType
 } from './data.js';
+import GameScreenViewDouble from '../game-screens/views/game-view-double.js';
+import GameScreenViewSingle from '../game-screens/views/game-view-single.js';
+import GameScreenViewTriple from '../game-screens/views/game-view-triple.js';
+
+const TYPE_PAINT = `paint`;
+
+const GameTypes = {
+  [GameType.DOUBLE]: GameScreenViewDouble,
+  [GameType.SINGLE]: GameScreenViewSingle,
+  [GameType.TRIPLE]: GameScreenViewTriple,
+};
 
 // Подсчет очков
 export const calculatePoints = (answers, lives) => {
@@ -67,7 +80,35 @@ export const getNextState = (state, answer) => {
   question += 1;
   state = changeLevel(state, question);
   lives = answer === AnswerValue.WRONG ? lives -= 1 : lives;
-  state = Object.assign({}, setLives(state, lives), {answers});
+  state = Object.assign({}, setLives(state, lives), {
+    answers
+  });
   state = setTime(state, Settings.TIME_FOR_QUESTION);
   return state;
+};
+
+export const getViewType = (state) => {
+  const viewType = GameTypes[testGame[state.question].type];
+  return viewType;
+};
+
+export const checkThirdGameTypeAnswer = (state) => {
+  const paintIndexArr = [];
+  const photoIndexArr = [];
+  testGame[state.question].answers.forEach((it, index) => {
+    if (it.answer === TYPE_PAINT) {
+      paintIndexArr.push(index);
+    } else {
+      photoIndexArr.push(index);
+    }
+  });
+  const currentIndex = paintIndexArr.length < photoIndexArr.length ? paintIndexArr[0] : photoIndexArr[0];
+  return currentIndex;
+};
+
+export const clearAllTimers = () => {
+  let maxId = setTimeout(function () {});
+  while (maxId--) {
+    clearTimeout(maxId);
+  }
 };
