+++ js/greeting/greeting-screen.js
